pyproject.toml
[project]
name = "docker-agent"
version = "0.1.0"
description = "Simple docker agent for observability"
readme = "README.md"
requires-python = "==3.12.12"
dependencies = [
    "docker>=7.1.0",
    "langchain[deepseek,openai]>=1.0.8",
    "langfuse[langchain]>=3.10.1",
    "langgraph>=1.0.3",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.6",
]
